Properly implement generic early returns.
