+++ chaos/errors.go
@@ -3,3 +3,4 @@ package chaos
 import "errors"
 
 var ErrChaoticBehavior = errors.New("chaotic behavior")
+var ErrEarlyReturn = errors.New("early return")
